Sign up

+++ src/Domain/Language/languageUseCases.ts
@@ -8,7 +8,9 @@ import {
   ResponseAPISuccess,
 } from "../Init/entities/response";
 import { Language, LanguageSchema, OrmLanguage, SearchLanguagesByTypeAndUser, SearchLanguagesByUser } from "./Entities/Language";
- 
+import { UserLanguage, UserLanguageSchema, OrmUserLanguage, } from "./Entities/UserLanguage";
+import { User, UserSchema, OrmUser } from "../Session/Entities/User";
+
 
 const ajv = new Ajv();
 
@@ -68,8 +70,8 @@ export default class LanguageManagement {
 
   getVersion = async (_req: Request, res: Response): Promise<Response> => {
     try {
-      let version={
-        Version: "1.4.23.01.27.08"
+      let version = {
+        Version: "1.4.23.02.22.14"
       }
 
       const responseApi = new ResponseAPISuccess();
@@ -144,17 +146,17 @@ export default class LanguageManagement {
       return res.status(500).json(responseApi);
     }
   };
-  
+
   getLanguageByTypeAndUser = async (
     req: Request,
     res: Response
   ): Promise<Response> => {
     try {
-      let type=req.params.type.toUpperCase() ;
+      let type = req.params.type.toUpperCase();
       console.log(type);
-      let userId=req.params.userId;
-      console.log(userId); 
-      const searchLanguageByType =(await  SearchLanguagesByTypeAndUser(type, userId));
+      let userId = req.params.userId;
+      console.log(userId);
+      const searchLanguageByType = (await SearchLanguagesByTypeAndUser(type, userId));
 
       const responseApi = new ResponseAPISuccess();
       responseApi.message = "Language were found successfully";
@@ -174,16 +176,16 @@ export default class LanguageManagement {
       return res.status(500).json(responseApi);
     }
   };
-  
+
   getLanguagesByUser = async (
     req: Request,
     res: Response
   ): Promise<Response> => {
     let transactionId = "getLanguageByUser_"
-    try { 
-      let userId=req.params.userId;
-      console.log(userId); 
-      const searchLanguageByType =(await  SearchLanguagesByUser(userId));
+    try {
+      let userId = req.params.userId;
+      console.log(userId);
+      const searchLanguageByType = (await SearchLanguagesByUser(userId));
 
       const responseApi = new ResponseAPISuccess();
       responseApi.message = "Language were found successfully";
@@ -212,21 +214,34 @@ export default class LanguageManagement {
       if (!valid)
         throw new Error(
           "Something wrong in request definition. Validate: " +
-            JSON.stringify(validate.errors)
+          JSON.stringify(validate.errors)
         );
 
       let language: Language = new Language();
       language = Object.assign(language, req.body.data);
+      language.stateAccept="PENDING";
 
       validate = ajv.compile(LanguageSchema);
       valid = validate(req.body.data);
       if (!valid)
         throw new Error(
           "Something wrong in data definition. Validate: " +
-            JSON.stringify(validate.errors)
+          JSON.stringify(validate.errors)
         );
 
-      let newLanguage = await OrmLanguage.create(language, {
+      let userId = req.body.user;
+
+      let user: any = (await OrmUser.findOne({
+        where: {
+          id: userId
+        },
+      }));
+
+      if (!user) {
+        throw "You are not authorized to create this record."
+      }
+
+      let newLanguage: any = await OrmLanguage.create(language, {
         fields: [
           "name",
           "abstractSyntax",
@@ -235,6 +250,17 @@ export default class LanguageManagement {
           "stateAccept",
           //add semantics support
           "semantics",
+        ],
+      });
+
+      let userLanguage: UserLanguage = new UserLanguage();
+      userLanguage.language_id = newLanguage.id;
+      userLanguage.user_id = user.id;
+
+      let newUserLanguage = await OrmUserLanguage.create(userLanguage, {
+        fields: [
+          "user_id",
+          "language_id",
         ],
       });
 
@@ -267,20 +293,34 @@ export default class LanguageManagement {
       if (!valid)
         throw new Error(
           "Something wrong in request definition. Validate: " +
-            JSON.stringify(validate.errors)
+          JSON.stringify(validate.errors)
         );
 
       let language: Language = new Language();
       language = Object.assign(language, req.body.data);
       language.id = parseInt(req.params.id);
+      language.stateAccept="PENDING";
 
       validate = ajv.compile(LanguageSchema);
       valid = validate(req.body.data);
       if (!valid)
         throw new Error(
           "Something wrong in data definition. Validate: " +
-            JSON.stringify(validate.errors)
+          JSON.stringify(validate.errors)
         );
+
+      const userId = req.body.user;
+
+      let userLanguage = (await OrmUserLanguage.findOne({
+        where: {
+          user_id: userId,
+          language_id: language.id,
+        },
+      })) as UserLanguage;
+
+      if (!userLanguage) {
+        throw "You are not authorized to modify this record."
+      }
 
       let updateLanguage = await OrmLanguage.update(
         {
@@ -322,6 +362,18 @@ export default class LanguageManagement {
   deleteLanguage = async (req: Request, res: Response) => {
     try {
       const id = parseInt(req.params.id);
+      const userId = req.params.userId;
+
+      let userLanguage = (await OrmUserLanguage.findOne({
+        where: {
+          user_id: userId,
+          language_id: id,
+        },
+      })) as UserLanguage;
+
+      if (!userLanguage) {
+        throw "You are not authorized to delete this record."
+      }
 
       const deleteLanguage = (await OrmLanguage.destroy({
         where: { id: id },
